requirements.txt

Django>=5.0,<6.0
django-environ>=0.11.2
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
Pillow>=10.4.0
